requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.12.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
requests==2.31.0
six==1.16.0
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
whisper==1.1.10
